examples/instructor-gemini-flash/requirements.txt
burr[start]
google-generativeai
instructor
pydantic
python-dotenv
